Move getOrganization from helpers to getters
