+++ eisbuk-admin/src/__tests__/customerDataTriggers.test.ts
@@ -5,7 +5,7 @@ import {
   OrgSubCollection,
 } from "eisbuk-shared";
 
-import { getOrganization } from "@/utils/helpers";
+import { getOrganization } from "@/lib/getters";
 import { adminDb } from "@/__testSettings__";
 
 import { deleteAll } from "@/__testUtils__/firestore";
