+++ eisbuk-admin/src/lib/getters.ts
@@ -1,0 +1,5 @@
+import { __organization__ } from "@/lib/constants";
+
+export const getOrganization = (): string => {
+  return localStorage.getItem("organization") || __organization__;
+};
