+++ eisbuk-admin/src/store/actions/__tests__/customerOperations.test.ts
@@ -9,7 +9,7 @@ import {
 
 import { Action, NotifVariant } from "@/enums/store";
 
-import { getOrganization } from "@/utils/helpers";
+import { getOrganization } from "@/lib/getters";
 import { adminDb } from "@/__testSettings__";
 
 import { NotificationMessage } from "@/enums/translations";
