+++ eisbuk-admin/src/store/actions/__tests__/slotOperations.test.ts
@@ -5,7 +5,7 @@ import {
   SlotType,
 } from "eisbuk-shared";
 
-import { getOrganization } from "@/utils/helpers";
+import { getOrganization } from "@/lib/getters";
 
 import { Action, NotifVariant } from "@/enums/store";
 import { NotificationMessage } from "@/enums/translations";
