+++ eisbuk-admin/src/store/actions/copyPaste.ts
@@ -9,7 +9,7 @@ import {
   luxon2ISODate,
 } from "eisbuk-shared";
 
-import { getOrganization } from "@/utils/helpers";
+import { getOrganization } from "@/lib/getters";
 
 import { Action } from "@/enums/store";
