+++ eisbuk-admin/src/themes.ts
@@ -3,7 +3,7 @@ import responsiveFontSizes from "@material-ui/core/styles/responsiveFontSizes";
 
 import * as colors from "@material-ui/core/colors";
 
-import { getOrganization } from "@/utils/helpers";
+import { getOrganization } from "@/lib/getters";
 
 const base = {
   palette: {
